requirements.txt
future==0.18.2
pillow==8.0.1
numpy<1.19.0,>=1.14.5 
setuptools>=41.0.0
torch==1.7.1
tensorboard>=1.15
torchtext==0.8.1
sacrebleu>=1.3.6
subword-nmt==0.3.7
matplotlib==3.4.0
seaborn==0.11.1
pyyaml>=5.1
pylint==2.7.2
six==1.12
wrapt==1.11.1
sacremoses==0.0.43
faiss-gpu==1.7.0
bytedkafka==0.0.17
tqdm==4.59.0
numba==0.53.1
npy-append-array==0.9.6
elasticsearch==7.12.0
spacy==3.0.6